LayoutTests/fast/css/font-systemFontID-parsing.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<div id="testDivInvalid" style="font: caption icon"></div>
<div id="testDivValid" style="font: caption"></div>
<script>
description("Tests 'font' CSS property parsing with system font values.");
shouldBeNull("testDivInvalid.style.getPropertyCSSValue('font')");
shouldBeEqualToString("testDivValid.style.getPropertyCSSValue('font')", "caption");
shouldBeNull("testDivInvalid.style.getPropertyCSSValue('font-family')");
shouldBeNull("testDivValid.style.getPropertyCSSValue('font-family')");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
